- fixed compile warning
git-svn-id: svn+ssh://svn.code.sf.net/p/opensips/svn/trunk@1373 689a6050-402a-0410-94f2-e92a70836424

+++ items.c
@@ -1192,7 +1192,7 @@ static int xl_get_color(struct sip_msg *msg, xl_value_t *res, xl_param_t *param,
         
 	if(param->val.s[0]!='_')
 	{
-		if (islower(param->val.s[0]))
+		if (islower((int)param->val.s[0]))
 		{
 			/* normal font */
 			append_sstring(p, end, "0;");
